+++ flashcard/src/flashcard.py
@@ -1,7 +1,7 @@
 from random import shuffle
-import platform
 from handlers.filereader import FileReader
 from data.card import Card
+
 
 class Flashcard:
     """Sovellustoteutuksen sisältävä luokka. Annetaan viitteenä käyttöliittymälle.
@@ -19,7 +19,6 @@ class Flashcard:
 
         self.__load_setting_file("setting.xml")
 
-
     def get_files_in_folder(self):
         """Hakee _packfolder-muuttujan osoittaman kansion .xmlpack-tiedostot
 
@@ -30,7 +29,6 @@ class Flashcard:
         self.__packs = FileReader.get_files(
             self.__packfolder, ".xmlpack", True)
         return self.__packs
-
 
     def load_pack(self, path):
         """Lataa tietyn korttipakan tiedot valmiiksi käsittelyä varten
@@ -49,7 +47,6 @@ class Flashcard:
         self.generate_pack_linear_order()
         return True
 
-
     def save_pack(self):
         """Tallentaa aktiivisen pakan xml-tiedostoksi. Tiedoston osoite on Pack.path.
 
@@ -62,7 +59,6 @@ class Flashcard:
         self.__activepack.unsaved_changes = False
         return FileReader.save_to_xml(self.__activepack)
 
-
     def generate_pack_linear_order(self):
         """Järjestää kortit normaaliin järjestykseen,
          jotta ne voidaan sitten nostaa yksitellen Flashcard.get_next_card()-metodilla.
@@ -70,7 +66,6 @@ class Flashcard:
 
         if self.__activepack:
             self.__packorder = list(range(len(self.__activepack)))
-
 
     def generate_pack_random_order(self):
         """Järjestää kortit satunnaiseen järjestykseen,
@@ -95,7 +90,6 @@ class Flashcard:
             raise IndexError("self.__activepack does not exist")
         return self.__activepack.name
 
-
     def get_sentences(self):
         """Palauttaa ladatun pakan korttien lauseet.
 
@@ -107,9 +101,9 @@ class Flashcard:
             return []
         return [self.__activepack.get_card(i).sentence for i in range(len(self.__activepack))]
 
-
     def get_next_card(self):
-        """Seuraten Flashcard.__packorder listaa, palauttaa seuraavan kortin käytössä olevasta pakasta.
+        """Seuraten Flashcard.__packorder listaa, palauttaa seuraavan kortin
+        käytössä olevasta pakasta.
 
         Returns:
             Card: __packorder-listan seuraavana ollut kortti.
@@ -120,7 +114,6 @@ class Flashcard:
             return None
         card = self.__activepack.get_card(self.__packorder.pop(-1))
         return card
-
 
     def get_pack_filename(self):
         """Palauttaa käytössä olevan pakan tiedostonimen
@@ -133,7 +126,6 @@ class Flashcard:
             return None
         return self.__activepack.path
 
-
     def get_pack_changed(self):
         """Kertoo, onko pakan sisältöön tehty muutoksia sitten viime tallennuksen.
 
@@ -144,7 +136,6 @@ class Flashcard:
         if self.__activepack is None:
             return False
         return self.__activepack.unsaved_changes
-
 
     def __load_setting_file(self, path):
         """Lataa ohjelman asetukset .xml-tiedostosta ja lisää puuttuvat asetukset .xml-tiedostoon
@@ -163,16 +154,19 @@ class Flashcard:
         changed = max(changed, self.__setting_set_if_none("font_family", ""))
         changed = max(changed, self.__setting_set_if_none("font_h1", "14"))
         changed = max(changed, self.__setting_set_if_none("font_p", "10"))
-        changed = max(changed, self.__setting_set_if_none("font_sentence", "12"))
-        changed = max(changed, self.__setting_set_if_none("font_reading", "11"))
-        changed = max(changed, self.__setting_set_if_none("font_translation", "11"))
+        changed = max(changed, self.__setting_set_if_none(
+            "font_sentence", "12"))
+        changed = max(changed, self.__setting_set_if_none(
+            "font_reading", "11"))
+        changed = max(changed, self.__setting_set_if_none(
+            "font_translation", "11"))
 
         if changed:
             FileReader.save_dict_to_file(self.__setting, path)
 
-
     def __setting_set_if_none(self, key, value):
-        """Lisää asetuksen sanakirjaan Flashcard.__setting, jos siellä ei ole jo samannimistä asetusta.
+        """Lisää asetuksen sanakirjaan Flashcard.__setting, jos
+        siellä ei ole jo samannimistä asetusta.
 
         Args:
             key (string): avain
@@ -187,7 +181,6 @@ class Flashcard:
             return True
         return False
 
-
     def get_setting(self, key):
         """Julkinen funktio, joka palauttaa asetuksen Flashcard.__setting-sanakirjasta
 
@@ -200,7 +193,6 @@ class Flashcard:
 
         return self.__setting[key]
 
-
     def set_setting(self, key, value):
         """Julkinen funktio, joka asettaa jonkin asetuksen Flashcard.__setting-sanakirjaan.
 
@@ -211,13 +203,11 @@ class Flashcard:
 
         self.__setting[key] = value
 
-
     def save_setting(self):
         """Julkinen funktio, joka tallentaa asetukset tiedostoon setting.xml.
         """
 
         FileReader.save_dict_to_file(self.__setting, "setting.xml")
-
 
     @property
     def file_error(self):
@@ -229,7 +219,6 @@ class Flashcard:
 
         return FileReader.lasterror
 
-
     @property
     def packfolder(self):
         """Palauttaa kansion osoitteen, mistä Flashcard-luokka etsii .xmlpack-tiedostoja.
@@ -240,20 +229,20 @@ class Flashcard:
 
         return self.__packfolder
 
-
     def __pack_set(self):
         """Luokan sisäinen metodi, joka ajetaan aina ennen korttien muokkaamista.
-         Tarkastaa, onko käytössä oleva pakka olemassa, ja laittaa muutosta ilmaisevan muuttujan todeksi.
+        Tarkastaa, onko käytössä oleva pakka olemassa, ja
+        laittaa muutosta ilmaisevan muuttujan todeksi.
 
         Returns:
-            boolean: True, jos pakka oli olemassa ja Pack.unsaved_changes on muutettu todeksi, muuten False.
+            boolean: True, jos pakka oli olemassa ja
+            Pack.unsaved_changes on muutettu todeksi, muuten False.
         """
 
         if self.__activepack is None:
             return False
         self.__activepack.unsaved_changes = True
         return True
-
 
     def new_card(self, sentence, reading, translation):
         """Lisää uuden kortin käytössä olevaan korttipakkaan.
@@ -277,7 +266,6 @@ class Flashcard:
 
         return True
 
-
     def get_card_sentence(self, index):
         """Hae tämänhetkisestä korttipakasta kortin japaninkielinen lause
 
@@ -293,7 +281,6 @@ class Flashcard:
             return None
         return self.__activepack.get_card(index).sentence
 
-
     def set_card_sentence(self, index, value):
         """Aseta tämänhetkisestä korttipakasta löytyvän kortin japaninkielinen lause.
 
@@ -305,7 +292,6 @@ class Flashcard:
         if self.__pack_set() is False:
             return
         self.__activepack.get_card(index).sentence = value
-
 
     def get_card_reading(self, index):
         """Hae tämänhetkisestä korttipakasta kortin lauseen lukutapa
@@ -322,12 +308,12 @@ class Flashcard:
             return None
         return self.__activepack.get_card(index).reading
 
-
     def set_card_reading(self, index, value):
         """Aseta tämänhetkisestä korttipakasta löytyvän kortin lukutapa.
 
         Args:
-            index (int): Kortin lauseen sijainti Flashcard.get_sentences()-metodin palauttamassa listassa.
+            index (int): Kortin lauseen sijainti Flashcard.get_sentences()-metodin
+            palauttamassa listassa.
             value (string): Lukutapa
         """
 
@@ -354,7 +340,8 @@ class Flashcard:
         """Aseta tämänhetkisestä korttipakasta löytyvän kortin käännöslause.
 
         Args:
-            index (int): Kortin lauseen sijainti Flashcard.get_sentences()-metodin palauttamassa listassa.
+            index (int): Kortin lauseen sijainti
+            Flashcard.get_sentences()-metodin palauttamassa listassa.
             value (string): Käännöslause
         """