+++ flashcard/src/handlers/filereader.py
@@ -3,13 +3,12 @@ import xml.etree.ElementTree as ET
 from data.pack import Pack
 from data.card import Card
 
-#   Tiedostojen lukemiseen tarkoitettu luokka    #
-#   -----------------------------------------    #
-# Tätä luokkaa käytetään erilaisten tiedostojen  #
-# ja hakemistojen lukemiseen                     #
-
-
 class FileReader:
+    """Tiedostojen lukemiseen ja tallentamiseen tarkoitettu luokka
+
+    Attributes:
+        lasterror: Viimeisin virheviesti lukiessa/tallentaessa tiedostoa
+    """
 
     lasterror = ""
 
@@ -17,6 +16,17 @@ class FileReader:
     # kansio luodaan, jos sitä ei ole ja "create" on tosi
     @staticmethod
     def get_files(path, suffix="", create=False):
+        """Palauttaa tiedostot, jotka päättyvät tiettyyn merkkijonoon
+
+        Args:
+            path (_type_): Kansion osoite.
+            suffix (str, optional): Palauta ainoastaan tiedostot, joiden
+            nimi päättyy tähän. Defaults to "".
+            create (bool, optional): Luo kansio, jos tosi ja kansiota ei löydy. Defaults to False.
+
+        Returns:
+            list: Tiedostojen nimet, jotka täyttävät annetut argumentit.
+        """
 
         if not os.path.exists(path):
             if create:
@@ -30,9 +40,17 @@ class FileReader:
 
         return files
 
-    # Lataa .xmlpack tiedosto
     @staticmethod
     def load_from_xml(path):
+        """Lataa .xmlpack-tiedosto Pack-luokkaan.
+
+        Args:
+            path (string): Tiedoston osoite.
+
+        Returns:
+            Pack: Luotu korttipakka.
+            None: Virhe lukiessa tiedostoa.
+        """
 
         if not os.path.exists(path):
             FileReader.lasterror = f'Tiedostoa "{path}" ei löytynyt'
@@ -57,19 +75,26 @@ class FileReader:
                 pack.add_card(
                     Card(p_sentence.text, p_reading.text, p_translation.text))
 
-        # TODO: Siirtyminen pois yleisestä Exception-luokasta
-        # Käytetään pylint disable, jottei lokissa ole samaa ongelmaa kahta kertaa
-        except Exception as ex:  # pylint: disable=broad-except
+
+        except ET.ParseError as ex:
             FileReader.lasterror = f'Virhe lukiessa tiedostoa "{path}]"\n' + str(
                 ex)
             return None
 
         return pack
 
-    # Tallenna .xmlpack tiedosto
-    # Funktio käyttää Pack.path muuttujaa osoitteena.
+
     @staticmethod
     def save_to_xml(pack):
+        """Tallenna korttipakka .xmlpack-tiedostoon.
+
+        Args:
+            pack (Pack): Korttipakka-objekti.
+
+        Returns:
+            boolean: Tosi, jos tallennus onnistui, muuten epätosi.
+        """
+
         # TODO: Kauniimman xml-tiedoston tuottaminen
         # Vasta Python 3.9 valitettavasti tarjoaa indent() komennon
         # https://stackoverflow.com/questions/3095434/inserting-newlines-in-xml-file-generated-via-xml-etree-elementtree-in-python
@@ -95,17 +120,26 @@ class FileReader:
             tree = ET.ElementTree(xmlpack)
             tree.write(pack.path, "UTF-8")
 
-        # TODO: Siirtyminen pois yleisestä Exception-luokasta
-        # Käytetään pylint disable, jottei lokissa ole samaa ongelmaa kahta kertaa
-        except Exception as ex:  # pylint: disable=broad-except
+        except ET.ParseError as ex:
             FileReader.lasterror = "Virhe kirjoittaessa XML-tiedostoa\n" + \
                 str(ex)
-            return None
+            return False
 
         return True
 
+
     @staticmethod
     def load_to_dict(path):
+        """Lataa kaikki xml-tiedoston elementit sanakirjaan.
+
+        Args:
+            path (string): Tiedoston osoite.
+
+        Returns:
+            dict: Sanakirja, sisältäen elementin ja sen tekstin.
+            None: Luku epäonnistui.
+        """
+
         if not os.path.exists(path):
             FileReader.lasterror = f'Tiedostoa "{path}" ei löytynyt'
             return None
@@ -122,28 +156,40 @@ class FileReader:
                 name = element.find('name')
                 value = element.find('value')
                 if name is None or value is None:
-                    FileReader.lasterror = f'Jätetään pois asetus indeksissä [{i}] palautettavasta listasta. Tiedosto: "{path}"'
+                    FileReader.lasterror = f"""Jätetään pois asetus indeksissä [{i}]
+                     palautettavasta listasta. Tiedosto: "{path}"."""
                     continue
 
                 name = name.text
                 value = value.text
 
                 if name in dictionary:
-                    FileReader.lasterror = f'Avain "{name}" löytyy jo listasta ({dict[name]} -> {value}). Tiedosto: "{path}"'
+                    FileReader.lasterror = f"""Avain "{name}" löytyy jo listasta
+                     ({dict[name]} -> {value}). Tiedosto: "{path}"."""
 
                 dictionary[name] = value
 
-        # TODO: Siirtyminen pois yleisestä Exception-luokasta
-        # Käytetään pylint disable, jottei lokissa ole samaa ongelmaa kahta kertaa
-        except Exception as ex:  # pylint: disable=broad-except
+        except ET.ParseError as ex:
             FileReader.lasterror = f'Virhe lukiessa tiedostoa "{path}]"\n' + str(
                 ex)
             return None
 
         return dictionary
 
+
     @staticmethod
     def save_dict_to_file(dictionary, file_path):
+        """Tallenna sanakirjan .xml-tiedostoon, jossa sanakirjan avain on elementin
+        nimi ja avaimen sisältö elementin teksti.
+
+        Args:
+            dictionary (dict): Tallennettava sanakirja.
+            file_path (string): Tiedoston osoite.
+
+        Returns:
+            boolean: Tosi, jos tallennus onnistui, muuten epätosi.
+        """
+
         path = file_path.split("/")
         folder = "./" + "/".join(path[:len(path)-1])
         if not os.path.exists(folder):
@@ -160,11 +206,9 @@ class FileReader:
             tree = ET.ElementTree(xmlpack)
             tree.write(file_path, "UTF-8")
 
-        # TODO: Siirtyminen pois yleisestä Exception-luokasta
-        # Käytetään pylint disable, jottei lokissa ole samaa ongelmaa kahta kertaa
-        except Exception as ex:  # pylint: disable=broad-except
+        except ET.ParseError as ex:
             FileReader.lasterror = "Virhe kirjoittaessa XML-tiedostoa\n" + \
                 str(ex)
-            return None
+            return False
 
         return True