+++ flashcard/src/tests/test_flashcard.py
@@ -3,7 +3,7 @@ import unittest
 import pytest  # pylint: disable=import-error
 
 from flashcard import Flashcard
-from data.pack import Pack
+from data.card import Card
 
 
 class TestFlashcard(unittest.TestCase):
@@ -60,3 +60,52 @@ class TestFlashcard(unittest.TestCase):
 
         val = self.flashcard.get_sentences()
         self.assertEqual(set(val), {"この**文**はテストです。", "アプリを**検索**"})
+    
+    def test_get_sentences_returns_empty_list_when_no_cards(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/nocards.xmlpack")
+        val = self.flashcard.get_sentences()
+        self.assertEqual(len(val), 0)
+
+    def test_save_pack_returns_false_when_no_pack(self):
+        self.assertFalse(self.flashcard.save_pack())
+
+    def test_get_pack_name_returns_name(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/working.xmlpack")
+
+        self.assertEqual(self.flashcard.get_pack_name(), "Testipakka")
+
+    def test_next_card_returns_none_when_no_pack(self):
+        self.assertEqual(self.flashcard.get_next_card(), None)
+
+    def test_next_card_returns_none_when_no_cards_remaining(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/nocards.xmlpack")
+        self.assertEqual(self.flashcard.get_next_card(), None)
+
+    def test_next_card_returns_card(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/working.xmlpack")
+        card = self.flashcard.get_next_card()
+        self.assertTrue(isinstance(card, Card))
+
+    def test_get_pack_filename_returns_filename(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/working.xmlpack")
+
+        self.assertEqual(self.flashcard.get_pack_filename(), "src/tests/test_files/working.xmlpack")
+
+    def test_get_pack_filename_returns_none_when_no_pack(self):
+        self.assertEqual(self.flashcard.get_pack_filename(), None)
+
+    def test_get_pack_changed_returns_changes(self):
+        self.flashcard.load_pack(
+            "src/tests/test_files/working.xmlpack")
+
+        self.assertFalse(self.flashcard.get_pack_changed())
+        self.flashcard.set_card_sentence(0, "a**a**a")
+        self.assertTrue(self.flashcard.get_pack_changed())
+
+    def test_get_pack_changed_returns_false_on_empty(self):
+        self.assertFalse(self.flashcard.get_pack_changed())