+++ flashcard/src/ui/tkinter.py
@@ -1,5 +1,6 @@
-from tkinter import Tk, Toplevel, Menu, constants, ttk, font
+from tkinter import Tk, Toplevel, Menu, constants, ttk, font, filedialog
 from ui.interfaces.open import TkinterGUIOpen
+from ui.interfaces.open import TkinterGUIPack
 
 
 class TkinterGUI:
@@ -22,21 +23,29 @@ class TkinterGUI:
 
         self.__reload_settings()
         self.init_menubar()
-    
+
+
     def __reload_settings(self):
         """Lataa käyttöliittymän asetukset, kuten fontit.
         """
 
         family = self.__service.get_setting("font_family")
 
-        self.font_h1 = font.Font(family=family, size=self.__service.get_setting("font_h1"), weight="bold")
-        self.font_p = font.Font(family=family, size=self.__service.get_setting("font_p"))
-        self.font_p_underline = font.Font(family=family, size=self.__service.get_setting("font_p"), underline=1)
+        self.font_h1 = font.Font(
+            family=family, size=self.__service.get_setting("font_h1"), weight="bold")
+        self.font_p = font.Font(
+            family=family, size=self.__service.get_setting("font_p"))
+        self.font_p_underline = font.Font(
+            family=family, size=self.__service.get_setting("font_p"), underline=1)
 
-        self.font_sentence = font.Font(family=family, size=self.__service.get_setting("font_sentence"))
-        self.font_sentence_highlight = font.Font(family=family, size=self.__service.get_setting("font_sentence"), weight="bold")
-        self.font_translation = font.Font(family=family, size=self.__service.get_setting("font_translation"))
-        self.font_reading = font.Font(family=family, size=self.__service.get_setting("font_reading"))
+        self.font_sentence = font.Font(
+            family=family, size=self.__service.get_setting("font_sentence"))
+        self.font_sentence_highlight = font.Font(
+            family=family, size=self.__service.get_setting("font_sentence"), weight="bold")
+        self.font_translation = font.Font(
+            family=family, size=self.__service.get_setting("font_translation"))
+        self.font_reading = font.Font(
+            family=family, size=self.__service.get_setting("font_reading"))
 
 
     def init_menubar(self):
@@ -46,9 +55,10 @@ class TkinterGUI:
         menu = Menu(self.__window)
         self.__window['menu'] = menu
 
-        menu.add_command(label='Avaa')
-        menu.entryconfigure('Avaa', state=constants.DISABLED)
+        menu.add_command(label='Avaa', command=self._popup_dialog)
+        #menu.entryconfigure('Avaa', state=constants.DISABLED)
         menu.add_command(label='Asetukset', command=self._popup_setting)
+
 
     def run(self):
         """Aloittaa käyttöliittymän toteutusloopin.
@@ -58,6 +68,7 @@ class TkinterGUI:
 
         self.__window.mainloop()
 
+
     def add_menu(self, menu):
         """Lisää valikko TkinterGUI.__menu-listaan.
 
@@ -65,9 +76,11 @@ class TkinterGUI:
             menu (TkinterGUITemplate): Valikko
         """
 
-        self.__menu[-1].hide()
+        if len(self.__menu) > 0:
+            self.__menu[-1].hide()
         self.__menu.append(menu)
         menu.show()
+
 
     def pop_menu(self):
         """Tuhoa viimeisin valikko listalta.
@@ -77,12 +90,36 @@ class TkinterGUI:
         self.__menu.pop(-1)
         self.__menu[-1].show()
 
+
+    def pop_all(self):
+        """Tuhoa kaikki valikot listalta."""
+        for menu in self.__menu:
+            menu.destroy()
+        self.__menu.clear()
+
+
     def __reload_menus(self):
         """Uudelleenalusta kaikki TkinterGUI.__menu-listan valikot.
         """
 
         for menu in self.__menu:
             menu.rebuild()
+    
+
+    def _popup_dialog(self):
+        """Avaa tiedostonvalitsemisikkuna, ja avaa tiedosto, jos sellainen valitaan.
+        """
+        
+        ftype = [('Packs', '*.xmlpack')]
+        path = filedialog.askopenfilename(filetypes=ftype)
+
+        if len(path) == 0:
+            return
+
+        self.pop_all()
+        self.add_menu(TkinterGUIOpen(self.__window, self.__service, self))
+        self.add_menu(TkinterGUIPack(self.__window, self.__service, self, path))
+
 
     def _popup_setting(self):
         """Asetukset-popup valikon alustava metodi. Tuhotaan vanha aina, kun avataan uudelleen.
@@ -93,7 +130,7 @@ class TkinterGUI:
                 val.destroy()
             self.__settings_objects.clear()
 
-        top= Toplevel(self.__window)
+        top = Toplevel(self.__window)
         top.grab_set()
         top.minsize(350, 400)
         top.maxsize(350, 400)
@@ -108,7 +145,8 @@ class TkinterGUI:
         frame.grid_rowconfigure(9, weight=1)
 
         # Fontin valinta
-        ttk.Label(frame, text="Fontti", font=self.font_p).grid(column=0, row=0, columnspan=2, sticky="W")
+        ttk.Label(frame, text="Fontti", font=self.font_p).grid(
+            column=0, row=0, columnspan=2, sticky="W")
 
         fonts = font.families()
         w_font_family = ttk.Combobox(frame, values=fonts)
@@ -122,40 +160,48 @@ class TkinterGUI:
         ttk.Label(frame).grid(row=2, column=0)
 
         # Fonttikoot
-        w_font_h1 = self.__setting_font_size_widget(frame, 
+        w_font_h1 = self.__setting_font_size_widget(frame,
             "Otsikon fonttikoko", 3, self.__service.get_setting('font_h1'), 10, 22)
-        w_font_p = self.__setting_font_size_widget(frame, 
+        w_font_p = self.__setting_font_size_widget(frame,
             "Tekstin fonttikoko", 4, self.__service.get_setting('font_p'), 6, 22)
 
         ttk.Label(frame).grid(row=5, column=0)
 
-        w_font_sentence = self.__setting_font_size_widget(frame, 
+        w_font_sentence = self.__setting_font_size_widget(frame,
             "Lauseen fonttikoko", 6, self.__service.get_setting('font_sentence'), 6, 22)
-        w_font_reading = self.__setting_font_size_widget(frame, 
+        w_font_reading = self.__setting_font_size_widget(frame,
             "Lukutavan fonttikoko", 7, self.__service.get_setting('font_reading'), 6, 22)
-        w_font_translation = self.__setting_font_size_widget(frame, 
+        w_font_translation = self.__setting_font_size_widget(frame,
             "Käännöksen fonttikoko", 8, self.__service.get_setting('font_translation'), 6, 22)
 
         # Tallennusnappi
-        w_save_button = ttk.Button(frame, text="Tallenna muutokset", command=self.__setting_save)
+        w_save_button = ttk.Button(
+            frame, text="Tallenna muutokset", command=self.__setting_save)
         w_save_button.grid(row=9, column=0, columnspan=2, sticky="S")
 
         # Lisätään tärkeät objektit listaan
-        self.__settings_objects = {"top":top, "w_save_button":w_save_button, "w_font_family":w_font_family, "w_font_h1":w_font_h1, "w_font_p":w_font_p,
-            "w_font_sentence":w_font_sentence, "w_font_reading":w_font_reading, "w_font_translation":w_font_translation}
+        self.__settings_objects = {"top": top, "w_save_button": w_save_button,
+            "w_font_family": w_font_family, "w_font_h1": w_font_h1, "w_font_p": w_font_p,
+            "w_font_sentence": w_font_sentence, "w_font_reading": w_font_reading, "w_font_translation": w_font_translation}
 
 
     def __setting_save(self):
         """Asetusikkunan tallenusmetodi, joka antaa tallennetut muutokset Flashcard-luokalle.
         """
-        
-        self.__service.set_setting("font_family", self.__settings_objects["w_font_family"].get())
 
-        self.__setting_save_numeric("font_h1", self.__settings_objects["w_font_h1"].get())
-        self.__setting_save_numeric("font_p", self.__settings_objects["w_font_p"].get())
-        self.__setting_save_numeric("font_sentence", self.__settings_objects["w_font_sentence"].get())
-        self.__setting_save_numeric("font_reading", self.__settings_objects["w_font_reading"].get())
-        self.__setting_save_numeric("font_translation", self.__settings_objects["w_font_translation"].get())
+        self.__service.set_setting(
+            "font_family", self.__settings_objects["w_font_family"].get())
+
+        self.__setting_save_numeric(
+            "font_h1", self.__settings_objects["w_font_h1"].get())
+        self.__setting_save_numeric(
+            "font_p", self.__settings_objects["w_font_p"].get())
+        self.__setting_save_numeric(
+            "font_sentence", self.__settings_objects["w_font_sentence"].get())
+        self.__setting_save_numeric(
+            "font_reading", self.__settings_objects["w_font_reading"].get())
+        self.__setting_save_numeric(
+            "font_translation", self.__settings_objects["w_font_translation"].get())
 
         self.__service.save_setting()
         self.__reload_settings()